daemon: split out snapctl support to its own api_*.go files
