+++ daemon/api.go
@@ -37,7 +37,6 @@ import (
 	"strings"
 
 	"github.com/gorilla/mux"
-	"github.com/jessevdk/go-flags"
 
 	"github.com/snapcore/snapd/asserts"
 	"github.com/snapcore/snapd/asserts/snapasserts"
@@ -52,7 +51,6 @@ import (
 	"github.com/snapcore/snapd/overlord/auth"
 	"github.com/snapcore/snapd/overlord/configstate"
 	"github.com/snapcore/snapd/overlord/configstate/config"
-	"github.com/snapcore/snapd/overlord/hookstate/ctlcmd"
 	"github.com/snapcore/snapd/overlord/servicestate"
 	"github.com/snapcore/snapd/overlord/snapshotstate"
 	"github.com/snapcore/snapd/overlord/snapstate"
@@ -151,12 +149,6 @@ var (
 	readyToBuyCmd = &Command{
 		Path: "/v2/buy/ready",
 		GET:  readyToBuy,
-	}
-
-	snapctlCmd = &Command{
-		Path:   "/v2/snapctl",
-		SnapOK: true,
-		POST:   runSnapctl,
 	}
 
 	sectionsCmd = &Command{
@@ -1580,83 +1572,6 @@ func readyToBuy(c *Command, r *http.Request, user *auth.UserState) Response {
 	}
 
 	return SyncResponse(true, nil)
-}
-
-var (
-	runSnapctlUcrednetGet = ucrednetGet
-	ctlcmdRun             = ctlcmd.Run
-)
-
-func runSnapctl(c *Command, r *http.Request, user *auth.UserState) Response {
-	var snapctlPostData client.SnapCtlPostData
-
-	if err := jsonutil.DecodeWithNumber(r.Body, &snapctlPostData); err != nil {
-		return BadRequest("cannot decode snapctl request: %s", err)
-	}
-
-	if len(snapctlPostData.Args) == 0 {
-		return BadRequest("snapctl cannot run without args")
-	}
-
-	_, uid, _, err := runSnapctlUcrednetGet(r.RemoteAddr)
-	if err != nil {
-		return Forbidden("cannot get remote user: %s", err)
-	}
-
-	// Ignore missing context error to allow 'snapctl -h' without a context;
-	// Actual context is validated later by get/set.
-	context, _ := c.d.overlord.HookManager().Context(snapctlPostData.ContextID)
-
-	// make the data read from stdin available for the hook
-	// TODO: use a forwarded stdin here
-	if snapctlPostData.Stdin != nil {
-		context.Lock()
-		context.Set("stdin", snapctlPostData.Stdin)
-		context.Unlock()
-	}
-
-	stdout, stderr, err := ctlcmdRun(context, snapctlPostData.Args, uid)
-	if err != nil {
-		if e, ok := err.(*ctlcmd.UnsuccessfulError); ok {
-			result := map[string]interface{}{
-				"stdout":    string(stdout),
-				"stderr":    string(stderr),
-				"exit-code": e.ExitCode,
-			}
-			return &resp{
-				Type: ResponseTypeError,
-				Result: &errorResult{
-					Message: e.Error(),
-					Kind:    client.ErrorKindUnsuccessful,
-					Value:   result,
-				},
-				Status: 200,
-			}
-		}
-		if e, ok := err.(*ctlcmd.ForbiddenCommandError); ok {
-			return Forbidden(e.Error())
-		}
-		if e, ok := err.(*flags.Error); ok && e.Type == flags.ErrHelp {
-			stdout = []byte(e.Error())
-		} else {
-			return BadRequest("error running snapctl: %s", err)
-		}
-	}
-
-	if context != nil && context.IsEphemeral() {
-		context.Lock()
-		defer context.Unlock()
-		if err := context.Done(); err != nil {
-			return BadRequest(i18n.G("set failed: %v"), err)
-		}
-	}
-
-	result := map[string]string{
-		"stdout": string(stdout),
-		"stderr": string(stderr),
-	}
-
-	return SyncResponse(result, nil)
 }
 
 // aliasAction is an action performed on aliases
