+++ daemon/export_api_snapctl_test.go
@@ -1,0 +1,32 @@
+// -*- Mode: Go; indent-tabs-mode: t -*-
+
+/*
+ * Copyright (C) 2018-2020 Canonical Ltd
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package daemon
+
+import (
+	"github.com/snapcore/snapd/overlord/hookstate"
+)
+
+func MockCtlcmdRun(mock func(*hookstate.Context, []string, uint32) ([]byte, []byte, error)) (restore func()) {
+	oldCtlcmdRun := ctlcmdRun
+	ctlcmdRun = mock
+	return func() {
+		ctlcmdRun = oldCtlcmdRun
+	}
+}
